requirements.txt
affine==2.3.0
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
billiard==3.6.3.0
boto3==1.12.39
botocore==1.15.39
celery==4.4.2
certifi==2020.4.5.1
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
dj-database-url==0.5.0
Django==3.0.5
django-colorful==1.3
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-leaflet==0.26.0
django-raster==0.8
docutils==0.15.2
Fiona==1.8.4
GDAL==2.3.3
geojson==2.5.0
gunicorn==20.0.4
importlib-metadata==1.6.0
jmespath==0.9.5
kombu==4.6.8
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
munch==2.5.0
numpy==1.18.2
Pillow==7.1.1
psycopg2==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
rasterio==1.0.21
rasterstats==0.14.0
s3transfer==0.3.3
Shapely==1.6.4.post1
simplejson==3.17.0
six==1.14.0
snuggs==1.4.7
sqlparse==0.3.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.0.1
wincertstore==0.2
zipp==3.1.0
